+++ spec/dummy/app/controllers/foobars_controller.rb
@@ -17,13 +17,14 @@ class FoobarsController < ApplicationController
 
   default_filter_by :renamed_foo_id, not_eq: 3
   default_order_by [{foo_id: :desc}]
-  query_includes_for :create, :index, are: [:foo]
+  query_includes :foo
+  #query_includes_for :create, :index, are: [:foo]
   query_includes_for :update, are: [:bar]
 
 private
 
   def foobar_params
-    params.require(:foobar).permit(:id, foo_attributes: [:id, :code])
+    params.require(:foobar).permit(:id, :foo_id)
 
   #rescue => e
   #  puts "Problem with foobar_params: #{params.inspect}"